backend/requirements.txt
<<<<<<< HEAD
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.2
=======
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.2
>>>>>>> c5ddd412e7966043910823035b0bfaf2b578edcf
pydantic-settings==2.1.0